requirements.txt
# Core dependencies for KAN-LNP framework
# Python >= 3.9 required

# Deep Learning
torch>=1.10.0
pykan>=0.2.8

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0

# Data Processing
pandas>=1.3.0
openpyxl>=3.0.0

# Machine Learning
scikit-learn>=1.0.0

# Hyperparameter Optimization
optuna>=3.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Symbolic Mathematics
sympy>=1.9
